+++ sdk/src/main/java/com/mapbox/maps/SnapshotterDestroyedException.kt
@@ -1,0 +1,6 @@
+package com.mapbox.maps
+
+/**
+ * Runtime exception thrown when the [Snapshotter] has already been destroyed.
+ */
+class SnapshotterDestroyedException : IllegalStateException("This snapshotter was already destroyed.")